refactor: gigantic commit, read desc...
~ I don't care the size of the commit, those are my dotfiles ~

In this commit I reorginized the things, cleaning up so there will be no
folder with lots of files alone and packages and all will be installed
in the right spots. That organization can be improved little by little
moving things around bit by bit, but for now it looks the way I want.