fix: using git submodules as flakes input
This is a temporary fix, because currently state of "?submodules=1" arg
is not working.

Why the arg is preferable than the flake input? because
now flake inputs can only use relative paths from the place the build
command is called, meaning that when we call it with `sudo`, it will
look for paths in `root` dir. Using absolute paths here is bad because
we explicitly write the username, which is not cool if anybody wants to
use another username.

I need to wait until flakes input allow relative path to look from top
level file location or the `submodules` arg to be fixed.