fix: patch niri depndency `smithay` (specifically `wayland-rs`)
**Why?**

On chromium (my main browser), but also brave and other chromium-based
browsers (reportedly electron apps too), a `data too big for buffer`
error was making the browser crash often. The patch is for `wayland-rs`
using a PR that is exposing `wl_display_set_default_max_buffer_size` and
reportedly fixes the issue in NIRI. While this thing is not merged, we
will rely on this patch for now and see if it actually fixes the issue.